Subject: front end flask container file updates-- main.py and accessory functions
main.py now imports accessory functions and calls a function to send data to an endpoint.
started creating a cache object for this server to handle all the requests and ignore data.
If person decides to look at a another query  in rapid fashion should be able to handle the overload of data.

TODO: find a way to make python asynchronous with response calls
Finish the cacheObject
Start making a logging decorator for our python functions.